Put coreg into source reconstruction scripts.
